+++ maven-resolver-transport-apache5x/src/main/java/org/eclipse/aether/transport/apache5x/ConnMgrConfig.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.transport.apache;
+package org.eclipse.aether.transport.apache5x;
 
 import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLContext;
