+++ maven-resolver-transport-apache5x/src/main/java/org/eclipse/aether/transport/apache5x/DeferredCredentialsProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.transport.apache;
+package org.eclipse.aether.transport.apache5x;
 
 import java.util.HashMap;
 import java.util.Iterator;
