+++ maven-resolver-transport-apache5x/src/main/java/org/eclipse/aether/transport/apache5x/LocalState.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.transport.apache;
+package org.eclipse.aether.transport.apache5x;
 
 import java.io.Closeable;
 import java.util.concurrent.ConcurrentHashMap;
