+++ maven-resolver-transport-apache5x/src/test/java/org/eclipse/aether/transport/apache5x/ApacheTransporterTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.transport.apache;
+package org.eclipse.aether.transport.apache5x;
 
 import java.io.File;
 import java.net.URI;
